feat: added isExists function in Delete function of user model to check does user exists

+++ Models/UserModel/UserModel.go
@@ -95,20 +95,16 @@ func Update(userId *int, userData *Entities.User) error {
 }
 
 func Delete(userId *int) error {
+	targetUserExists := isExists(userId);
+	if !targetUserExists {
+		return sql.ErrNoRows
+	}
+
 	query := `DELETE FROM users WHERE user_id = ?`;
 
-	result, err := DB.Connection.Exec(query, userId);
+	_, err := DB.Connection.Exec(query, userId);
 	if err != nil {
 		return err;
-	}
-
-	rowsAffected, err := result.RowsAffected();
-	if err != nil {
-		return err;
-	}
-
-	if rowsAffected == 0 {
-		return sql.ErrNoRows;
 	}
 
 	return nil;
